# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Spend Little",
        "description": "Essa API foi desenvolvida para manipular Spend Little",
        "contact": {
            "email": "gezerramo@gmail.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://spend-little.herokuapp.com/api/v1",
            "description": "API(Produção)"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API(Produção)"
        }
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "/user"
                ],
                "summary": "User",
                "description": "Rota responsavel por listar todos os usuarios!",
                "operationId": "29fe6261905366f57d192d4f07905f87",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "/user"
                ],
                "summary": "user",
                "description": "Rota responsavel por criar usuário!",
                "operationId": "9356d67ef616680f978450c4adf60a00",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "level_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "level_id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Fulano de Tal",
                                    "email": "fulano@gmail.com",
                                    "password": "123456",
                                    "level_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "/user"
                ],
                "summary": "User",
                "description": "Rota responsavel recuperar usuario!",
                "operationId": "4826814dc6639c07a116bb14cc3fced1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter example",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "/user"
                ],
                "summary": "User",
                "description": "Rota responsavel por editar usuario!",
                "operationId": "418e2b1abf275dae06ef49c5efdbbe98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter example",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "level_id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Fulano de Tal",
                                    "email": "fulano@gmail.com",
                                    "level_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}